+++ database/migrations/2024_09_12_000010_create_pickings_table.php
@@ -19,7 +19,7 @@ return new class () extends Migration {
         Schema::create('pickings', function (Blueprint $table) {
             $table->increments('id');
             $table = $this->groupOrgRelationship($table);
-            $table->unsignedInteger('shop_id')->index();
+            $table->unsignedSmallInteger('shop_id')->index();
             $table->foreign('shop_id')->references('id')->on('shops');
             $table->unsignedInteger('delivery_note_id')->index();
             $table->foreign('delivery_note_id')->references('id')->on('delivery_notes');
